+++ lib/fastlane/plugin/wpmreleasetoolkit/actions/android/an_localize_libs_action.rb
@@ -32,6 +32,13 @@ module Fastlane
       end
 
       def self.available_options
+        libs_hash_description = <<~KEYS
+          - `:library`: The library display name
+          - `:strings_path`: The path to the `strings.xml` file of the library
+          - `:exclusions`: An optional `Array` of string keys to exclude from merging
+          - `:source_id`: An optional `String` which will be added as the `a8c-src-lib` XML attribute
+            to strings coming from this library, to help identify their source in the merged file.
+        KEYS
         [
           FastlaneCore::ConfigItem.new(key: :app_strings_path,
                                        description: 'The path of the main strings file',
@@ -41,10 +48,7 @@ module Fastlane
           # See `Fastlane::Helper::Android::LocalizeHelper.merge_lib`'s YARD doc for more details on the keys expected for each Hash.
           FastlaneCore::ConfigItem.new(key: :libs_strings_path,
                                        env_name: 'LOCALIZE_LIBS_STRINGS_PATH',
-                                       description: 'The list of libs to merge. ' \
-                                         + 'Each item in the provided array must be a Hash with the keys `:library` (The library display name),' \
-                                         + '`:strings_path` (The path to the `strings.xml` file of the library) and ' \
-                                         + '`:exclusions` (Array of string keys to exclude from merging)',
+                                       description: "The list of libs to merge. Each item in the provided array must be a Hash with the following keys:\n#{libs_hash_description}",
                                        optional: false,
                                        type: Array),
         ]